Fix ObjectInspector crash under x64
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2704 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/ObjectInspector/Preference.cpp
@@ -739,7 +739,7 @@ void CPreference::CreateListItem(CArray<CListItem,CListItem&> &ListItems,
 	ListItem.strToolTip=(*pTableRow)[4];
 	ListItem.bIsDefault=bIsDefault;
 	ListItem.bIsDifferentValue=!bSameValueForAll;
-	ListItem.dwKeyValue=(DWORD)nKey;
+	ListItem.dwKeyValue = nKey;
 	ListItem.dwUserData=0;
 
 	CString strPrefName=(*pTableRow)[1];
